Separate reference gathering from rendering
This separates and refactors the process of gathering references from
the process of rendering Markdown. Separating the gathering process from
the rendering process in turn means we can decouple this from the
html-pipeline Gem. This in turn opens up the path for doing batch
processing when gathering references instead of processing everyting on
a per object basis.

The new namespace used for reference gathering is
Banzai::ReferenceParser. Each class in this namespace is tasked with
extracting a specific type of reference (e.g. users, commits, etc).