Batch reference extraction of Markdown documents
This adds support for extracting references from a set of documents in a
single pass. This in turn means that reference parsers can often
retrieve all references using a single query, with an exception here and
there.

As an example, say you have 200 comments each mentioning a different
username. Before this commit every link to the user would be processed
individually leading to at least 200 SQL queries being performed. By
using the batch setup introduced in this commit we can reduce this down
to just a single SQL query.

Checking whether a user can see/reference a link is still done on a
per-reference basis, this will be addressed in following commits.