Pass Markdown authors to the "analyze" method
When rendering multiple documents it's not unlikely for these documents
to have different authors. Previously this was not an issue as the
reference extraction pipeline was executed for every individual object,
passing the author to the #initialize method of the reference extractor
classes.

Now that documents are rendered in batches this is no longer possible.
To work around this the author is now passed to the "analyze" method and
its up to the Markdown pipeline to add any extra data to the rendered
HTML document. In this case the UserReferenceFilter class adds a
"data-author" attribute to links used for `@all` references. This in
turn allows the reference parsers to see if the author of the document
has the permissions required to actually reference everybody involved.